reorgnize code
